Fork cfg-expr
The direct event that triggered this fork was the MSRV bump, but I have
long thought that the cfg-expr API is complex for our use case. It is
still undecided whether we will upstream some changes in the future and
use cfg-expr again, replace it with an entirely new implementation, or
keep the fork.